Platform Check fixed Added World1 with 1 Obstacle Started Goomba

+++ app/src/main/java/eecs40/supermariobros/Goomba.java
@@ -10,7 +10,7 @@ public class Goomba extends Sprite {
 
     Bitmap bitmapGoomba;
 
-    public Goomba(float x, float y){
+    public Goomba(int x, int y){
         //Load images
 
         //Design hitbox
